WW-5459 Deprecate and repackage ActionChainResult
