+++ core/src/main/java/com/opensymphony/xwork2/interceptor/ChainingInterceptor.java
@@ -18,7 +18,6 @@
  */
 package com.opensymphony.xwork2.interceptor;
 
-import com.opensymphony.xwork2.ActionChainResult;
 import com.opensymphony.xwork2.ActionInvocation;
 import com.opensymphony.xwork2.Result;
 import com.opensymphony.xwork2.inject.Inject;
@@ -31,6 +30,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.struts2.StrutsConstants;
 import org.apache.struts2.Unchainable;
+import org.apache.struts2.result.ActionChainResult;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -117,7 +117,7 @@ import java.util.Map;
  *
  * @author mrdon
  * @author tm_jee ( tm_jee(at)yahoo.co.uk )
- * @see com.opensymphony.xwork2.ActionChainResult
+ * @see ActionChainResult
  *
  * @deprecated since 6.7.0, use {@link org.apache.struts2.interceptor.ChainingInterceptor} instead.
  */
