+++ core/src/test/java/com/opensymphony/xwork2/config/providers/ConfigurationProviderOgnlAllowlistTest.java
@@ -52,6 +52,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionSupport"),
+                Class.forName("org.apache.struts2.result.ActionChainResult"),
                 Class.forName("com.opensymphony.xwork2.ActionChainResult"),
                 Class.forName("com.opensymphony.xwork2.TextProvider"),
                 Class.forName("org.apache.struts2.interceptor.NoOpInterceptor"),
@@ -117,6 +118,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionSupport"),
+                Class.forName("org.apache.struts2.result.ActionChainResult"),
                 Class.forName("com.opensymphony.xwork2.ActionChainResult"),
                 Class.forName("com.opensymphony.xwork2.TextProvider"),
                 Class.forName("org.apache.struts2.interceptor.NoOpInterceptor"),
